modularizes dispatchAction into two functions, adds additional event listeners and emitters, refactors chatroom.js
